ignored old doc and archives format, more robust body extractor.
