+++ src/data_manager/collectors/scrapers/utils.py
@@ -1,8 +1,17 @@
 from scrapy.http import Response
 
-_IMAGE_EXTS = frozenset({
+IMAGE_EXTENSIONS = [
     ".png", ".jpg", ".jpeg", ".gif", ".bmp", ".svg", ".ico", ".webp"
-})
+]
+
+# .pdf, docs , xlsx, pptx are first class supported by MarkItDown
+IGNORED_DOCUMENT_EXTENSIONS = [
+    ".doc",
+    ".xls",
+    ".ppt",
+    ".zip",
+    ".rar",
+]
 
 def get_content_type(response: Response) -> str:
     """Decode the Content-Type header bytes to str."""
